pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "jarvis"
version = "1.0.0"
description = "Local AI voice assistant for Linux"
requires-python = ">=3.12"
dependencies = [
    "faster-whisper>=1.0.0",
    "openwakeword>=0.6.0",
    "piper-tts>=1.2.0",
    "pyaudio>=0.2.14",
    "sounddevice>=0.4.6",
    "anthropic>=0.34.0",
    "ollama>=0.3.0",
    "google-genai>=0.7.0",
    "PyQt6>=6.6.0",
    "PyQtGraph>=0.13.0",
    "pyyaml>=6.0",
    "numpy>=1.24.0",
    "loguru>=0.7.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "respx>=0.21.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
